feat: home creating and redirecting to list space
